fix: move tests to correct location
