+++ tests/test_KoalaBot.py
@@ -53,51 +53,6 @@ async def setup_clean_messages():
     yield dpytest
 
 
-@mock.patch("os.name", "posix")
-@mock.patch("sys.argv", ["KoalaBot.py", "--config", "/config/"])
-def test_get_config_from_argv_linux():
-    assert KoalaBot.get_arg_config_path() == "/config"
-
-
-@mock.patch("os.name", "posix")
-@mock.patch("sys.argv", ["KoalaBot.py", "--config", "/config"])
-def test_get_config_from_argv_linux_partial():
-    assert KoalaBot.get_arg_config_path() == "/config"
-
-@mock.patch("os.name", "posix")
-@mock.patch("sys.argv", ["KoalaBot.py", "--config", ""])
-def test_get_config_from_argv_linux_empty():
-    assert KoalaBot.get_arg_config_path() == "."
-
-
-@mock.patch("os.name", "nt")
-@mock.patch("sys.argv", ["KoalaBot.py", "--config", "/config/"])
-@mock.patch("os.getcwd", mock.MagicMock(return_value="C:/"))
-def test_get_config_from_argv_windows_relative():
-    assert KoalaBot.get_arg_config_path() == "C:\\config"
-
-
-@mock.patch("os.name", "nt")
-@mock.patch("sys.argv", ["KoalaBot.py", "--config", "/config"])
-@mock.patch("os.getcwd", mock.MagicMock(return_value="C:/"))
-def test_get_config_from_argv_windows_relative_partial():
-    assert KoalaBot.get_arg_config_path() == "C:\\config"
-
-
-@mock.patch("os.name", "nt")
-@mock.patch("sys.argv", ["KoalaBot.py", "--config", "\\config\\"])
-@mock.patch("os.getcwd", mock.MagicMock(return_value="C:/"))
-def test_get_config_from_argv_windows_relative_backslash():
-    assert KoalaBot.get_arg_config_path() == "C:\\config"
-
-
-@mock.patch("os.name", "nt")
-@mock.patch("sys.argv", ["KoalaBot.py", "--config", "D:/test/config/"])
-@mock.patch("os.getcwd", mock.MagicMock(return_value="C:/"))
-def test_get_config_from_argv_windows_absolute():
-    assert KoalaBot.get_arg_config_path() == "D:\\test\\config"
-
-
 def test_test_user_is_owner(test_ctx):
     assert KoalaBot.is_owner(test_ctx)
